[Chore] date 타입 Double -> Int64로 변경

+++ SoSoHappy/Sources/Network/Models/DTO/Request/FindFeedRequest.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct FindFeedRequest: Codable {
-    let date: Double
+    let date: Int64
     let nickName: String
 }
